ai_acne_analyzer/acne_model.py
from PIL import Image
import torch  # Example using PyTorch

def analyze_acne(image):
    # Load the image and preprocess it
    img = Image.open(image)
    processed_img = preprocess_image(img)  # Preprocessing step

    # Load your acne detection model (could be a pre-trained CNN model)
    model = load_model()

    # Get predictions from the model
    acne_severity, acne_type = model.predict(processed_img)

    # Example response with acne severity and type
    return {
        'acne_severity': acne_severity,
        'acne_type': acne_type,
        'recommendation': generate_recommendation(acne_severity)  # Optional
    }

def preprocess_image(img):
    # Code to resize/normalize image before passing to model
    return img  # Placeholder

def load_model():
    # Load your pre-trained acne detection model
    model = torch.load('acne_model.pth')  # Example using PyTorch
    model.eval()
    return model
